patch missing gc call for cert aggregator and remove unused code
