+++ bin/telcoin-network/Cargo.toml
@@ -41,6 +41,8 @@ rayon = { workspace = true }
 rpassword = { workspace = true }
 
 [dev-dependencies]
+tn-types = { workspace = true, features = ["test-utils"] }
+tn-reth = { workspace = true, features = ["test-utils"] }
 reth = { workspace = true }
 tempfile = { workspace = true }
 serde_json = { workspace = true }
@@ -62,6 +64,7 @@ k256 = "0.13.3"
 tonic = { workspace = true }
 nix = { version = "0.29", features = ["signal"] }
 cfg-if = { workspace = true }
+futures = { workspace = true }
 
 [features]
 default = []
