--- crates/consensus/.assets/diagram.drawio
@@ -1,1 +1,0 @@
-<mxfile host="app.diagrams.net" modified="2021-06-29T11:40:36.080Z" agent="5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.114 Safari/537.36" etag="OybOxsFT5Ujje8oWMdfj" version="14.8.2" type="device" pages="2"><diagram id="04SbDpS5iBzByXe31BT4" name="worker">7V1td5s2FP41Oaf7kB7ewR+btGl3tm7Z0nXrR2Jkw4KRC3Lj7NdP4sU2kmwrCaBL4pzT2ghZxs9zr+6LLuLMvlysP+bhMv6MI5SeWUa0PrPfn1mW6VgefWEtD1WLb/tVwzxPorrTtuEm+Q/VjUbdukoiVLQ6EoxTkizbjVOcZWhKWm1hnuP7drcZTtvfugznSGi4mYap2Pp3EpG4ag1cY9v+CSXzuPlm06jPLMKmc91QxGGE73ea7A9n9mWOManeLdaXKGXgNbhUn7vac3ZzYTnKiMoH7rJ/8C/J1fxrnnz4+umPq/Off7k+Nx23GudHmK7qn1xfLnloMMjxKosQG8Y4sy/u44Sgm2U4ZWfvKeu0LSaLlB6Z9G1BcnyHLnGK8/LTtlH+0TNhmswz2paiGb3gi1mSpk23DGd0tAvxJ9W/8gfKCVrvNNU/8SPCC0TyB9qlPuvUaDfi1hzfb8mzG0biXeK8ujGsBWa+GXqLKX1Tw/oYiA1bArGXMgyWLaC97ysmDBdpkqHz5uLe0S6mTS/O3Xag7+b1a1q9Xi35tthsWn5Fc5RFTfttLvn0tvPOlR36Dk5CKDnkkBjU/EooL6ggJdmcNrjboy+YCtX7c2ufuGEqD7O0VKY4iSJEpeqiLaScrHUgWB4nWK4oWFYgESynL7kKLEGs/kRpEt5SUizjhlKOcpH0NwTTs5jE7KRxj/M7lBc/PU/lJbQyXBM6i76rebjFhOAFPRGFRVyOa3ZDy0aZa15cT8KLI1N4ozdiHIGYm2SxfFW0OPBYEQ3dq2NlAlBZvCfTsmKkLPNkEeYPIyYl4EgJDJEU03KHJMUJBDhRRN3i+rBGqI3wv6vFsnbfPWaicxLjOc7C9FdcWnOG1r+IkIe6U7iiLLZIoBDmD/+w4d66zeG3evTy4P26dfRQH1WuRuOdW+z8OiHVQJZbH35rvoS+347DDpph9lJZUEmbokMeXtWPhPkckQP97FrWGZYHJSNHaUiSH+0QpHuWJyeW+2DZB8Vy46eeWO6WZQcWy+ZLYdmHxLIFiuTGtRDzCLeNa3QRkmlM+3wO72Se004wv/nIm7Ao0IJGjQVDZM3+T7LSw7plg6GDDq953LfqwEVyjbaL5Ej8VlOW1ektqeOKWbJR6dvTVcdSnSBBqY6YKxFU548VzleLv0MqzLlMTyRxyD3tzPRlhlkoEtJ/38tBGD5MgcLpXbGJVEptAqBLARcDBtp1ydahS13qhK1qTmApxf689Ebor3M8RUWBVW1JzIJjJvMhyzhTSnFeHsIxJZz4+9Joe1Dx1xJsS1ywJxqXp6uNo6g2JiitEbO7P/+mXawdbnnClnhIssReb6sTvh4HaRNGqEcREJwqX9WAwMot+Fosdx9R5xBBpzLJsFILvnMiuQeSYeWC/eOu4CeULpVzCjlapkmVUdhkEsof832FCgLAD5wE0FIKwSv0xvm1L/0sOKJJO01tFTQTxanNnoCa2pyT/eqBZAdYUOaeSO6BZAMUyZPjFvLmIZvGOc4oY+quSuWT0JEXSVGwYkRAFtIEt/ThjXhBuRopescKw7fXSluuEgZC+Qm1+p7jS4yGopKZFigl80a8kgyRXljBvKcnLfdi2TVhuUHNdR8ykX+iKaIXomoeF2GSEfqvXMjMWFnml8vrs80NNwnOGMZlgeyUhvwZ0W8z+XIBS3ITyMBrPFpK7SCs8TQKMrJFnuayYa3yeBawVR7Xf7VybY1Tri2Icu1Ck2tT4a5I0RVpOR87CLXlf1dBtsL/YdvKa8gjFhqPStkOoK4E0KZNWRjrb7jG1Ek421tkYQccUZV61Z/aciUM5BlHBqrUTxioJH3zs5/hLute4PS1LUmbnuIEF4Ca4Dwt2d5uEno7vFtDmTFVlmEVHpjiXWMAzFgArQjHG/EdPkNG+6qL9iaslS3/lIntll5Ypsw/ZWK7pRdWyY2pUHPz8nN1E3C5Og9eTmMwbzAYpTfoQamgf57P/7hYbwCWYZlDUyyQA+Dzmwa03JWnpRRmvNqgWgwDSxs870WwPJhlU2UZlmVrLhvWnMdvI6R9ymuu55Suf1K6nq+Ff3K63jQ05+sdLbNidyFzSySPxc9Pnw2bjsfvUrFATYebqepVh8x8Saj+kNnREjJDuHGzUZCjmgQrc9xcNiy/wnSA+RX+qOT6UcajUyVQ3f/CBmZOlDbAKPeXvKzMgfKddwVFe3OnHX2Nknl9j8Hj9q7Ucxue/k0xLHGXhd//+qIdqAm/r61k85xhIx9xIRwCTibvo+gHSlxtAQEUHzHpB0p2H/IplhZR3hcCc0UgHr9Lr3owHRwZqedgurHWwFSGr7LZGC9dGmPDtFU2NFtlw7RV/N0i+nGCaaoccDidLNVzLBW/ZZg7mbyd7PwFTzNbDr+/5OFh+7Zh4gIKBF3iodetSs3AwGCyLGg4iQloCDiZ4HASXcfT1PyIIMJvz5mu28dMbU6e8y09T9w+lDtmB92YRbUyq/uNhPesDHMxhOd1I3l80ZRv6ZS1YH8l7gyX1y0+o4+dOC9KOSqf0Oct1/ufz7fdQy/MovLRRgtUFOG83IxtlmO2YXgrJXxo/3HaVl1V58/iax6aN6Vyi2i78AykRRJF5QwrM2ltXez8KUltgTFlu5Z7Q5q4YP9+fX0JTbUAXS4i5GFWhOXydHGSFSFShScs+7euGnCGaT/H7iQ3whwDTm4mMrnh3aI0TZbFPsR2KAqLZfWE5FmyZrCJXOzytfvk3g6w9dvQBrYEWgmyZl/IbhaoBnjw8T5YBfx7wNlwdeMsq9nicY7x4nZVjBVj19GMsXTbp/HLssulyaRlCcMCrSLM4wM6UNlkcligFRJIIwSa90wBAC2rgRqlhyHMFpJ14YGxFReGR4qtMEHIPONhsVXZoWkU2E48cHOCuFA/UmzBTbeSej3Rro3LG3YUqrYHBlmWzx03yPwKn36QmwD+BYEcgJNk++UFz7wrEWgP6iR1jmMHmd9VxdUvyQq+8MhAtuGBrOAUjwxkvrJFP8YqzsVj6n/243zMU74q/zpyMILj03IgAdruD2ggu/rsBbfv4guBEjd467YHUS244AsOg2ZD+oEqLExJgSHBdwm7F46ExZ3A9PMXGFM0I89bXuxi9lJajpGoldWbWklKGBsmpnGYZSh9NWQ4E+1kiK7nb4iwBfqXSsIEHgdiBvyG4Dykk/sr4cDSPymJ0cHl9V+vBX9pBNwN/vQwx6xiaGvN6Y+KP+MIsR7/Aw==</diagram><diagram id="u77JRGBOouUTT-4SinIf" name="primary">7V1bc5s4FP41mdl9SIc7+DF1rt2mcS5Nt3nZUYxsaDEigGO7v34lDDYIGWOHi3zJTKdGCFk+37nrSJzI3dH0ygeedYtM6JxIgjk9kc9PJElUJA3/R1pm8xZd1ucNQ982407Lhkf7D4wbhbh1bJswyHQMEXJC28s29pHrwn6YaQO+jybZbgPkZL/VA0OYa3jsAyff+sM2Q2veaqjCsv0a2kMr+WZRiO+MQNI5bggsYKJJqkm+OJG7PkLh/NNo2oUOIV5Cl+uzM09RrUHn6f0evmhnp4M7/XQ+2OUmjyx+gg/dcOuhPdiDZw9I7frnLzcP0gvU3x9PxXjsd+CMY4LFPzacJRT00dg1IRlFOJE/Tyw7hI8e6JO7E8wzuM0KRw6+EvHHIPTRb9hFDvKjp2Uh+sN3gGMPXdzmwAH+EZ8HtuMk3Vzk4tE+l/yZyZShH8JpCuT4Z19BNIKhP8Nd4rtKDGDCwQmgkyU/KEbcZqV5QYsbQcyDw8XQSzrjDzGp2WR/+65dK5//uVP1d/0qhJeTzs3NqSTnqAxNzLbxZUyPLOF/jUdeLF4avkR+aKEhcoHzFSEvpv4vGIazuBMYh4jGBvjhGZGp5XdEbZc2mX70LdA1kx59BwSB3Z83xl2WECfCJEX3MVH+JSN8UpPLn4sB8cX5NHM1S66mdph6DF/9jL+DfF4+RC6SZzZjkgCN/X5M4dPra+fszbd+9tRvxkC46l1N7FhmBEyFIYzHewfPvzrSQH/583Q3/WKJ/33u3JzGgBGUClnOhw4I7fes8vkI/xTNOiW1N98K5FZcL7d1yZ4oZYVPMhjCx5A9pQLRY5JOzNGpAcnbVkCyoiZm5TOe61I4yRMmCKwF5hXLipKXFWY/jStZyRu0I+AVAy5wBXgrtvWwAFe5Alxh+LCaQ5zMV/xhSD48wD7E8/AXN/zkTtKCv3nR+68RsN0Q/wvwsAQDSXjq9pZBio1cQnUynNB3bEz7vzm1vlrW+i6uU9Z3YaHT5rc+z1c9er58eL7qTnq+6s54vgtPtwXPly16ylH0+BA9bSdFT9sZ0RNbFD0m6XSGh0KLYs7tyzh6KaplJTQtwkvxvFi20jKcl6TNYFjLjSkiqwwiJ22lmTb+hh7CPllKvVJZPVmgwJtLYfxUOi9KD6StGWguprmBIkZY/OztecNYzxu+hUav42C9QK3OvqbTraysbB9jj/3ZGmVSWC+TIisJW5s51PbVHOq7Zg87Je1hWQ1UuT1kMpDcSiqxPgbagg9STCdtxHQtMZDOlUPV2RmHyiihvJvN4gs5Qh2TfCWFRZTz0sLuyFeaT2xF3x4Y5h2+ID8u3jQg5gpfmMuMgOiYzo/MsMJdOt84+sDbBl5Vyztj8W4XoqhWPDleOajVKKo0B/EVRon59V9e46iOxtuakH6UPl70d9kFWb70t9Q5chAv+rssB1Wuv9krKx2ZWlnRm11ZEXenWMBo0TCwaZdf7c3L+XHJcjN5kCpasuysW/usW7BKrGfvw5qlXqKOoK41y/G927EnTy/6xRu4Hk5urWnQY9T9J/mOAYoYZAmA9jZGyY3TIJKlM9xB1Lzp8iadNVmkTK6BazokUTKCQQCGkGROBj4akQTKmORJJsj/Df2AlW5Jtc1nlTRTHIJBCItYIdYpjI1COQYgkNp94JzFN0a2aUaKhMV7WWX0gRTcBovfdOJGYDCSVpN+Z3ISa99X45wUWlHOzfPtEfBteOSmks6C1CI3sYOAlvP0Gad7mbRfmadPBQ/L0CAdP1DJ+83jhyRiSM2ohmoKpndfdktba3Fk0ayLsv6PWBxBOPaJNnmE/ruNyVBuASDATwaRMAATWy1O0/satcous9yNRtP7e5qbLR3ZVyyZyf73lGRCqXf3OsP2OdRf4Lj3ZXDt95Ita5xkePaszCkxGgsLslGRHO+sU7lSZ0egmppVVQqtgqqLQNkOR6v+RpZ1Ukm/Mg7HQWq01nYyMLmnlY0wH8wsb1iJUp+HWz8HMftVvqk8ehT/fDBLdfCIwgpW6z069FI1geLG+YiVKrtWCtW3s5MNrMtITZSpf4hVDa6UXSsVOitCcw4C82SVl2cGaq0gj+3/t1xrzZX+EXnnndZqtotmXZTV6SK/bBrHinLJqUTOidQlo6MQxh/7JHU7sPsAt3Ca5dHp/YZNZnmYIEnrQboC/is5aE8SushxYD9EfknMxp4JIniE0CLPXxGYIiA4xYfOwumt41OiILoHZg4CJu61YWl0gJFMofMKwr6F/zftIQxCXkWIPtak/USpeKyDrtjeMWpR2ZRvzeCxp3Msh2uLEfjymqU9XTvZAUZoLdlYNO0i630debO40w9gh6VNtw/fxsRI4+dGNobRHSYGPDLo2FfeDYeYLovTmtzVVBRDFUYtS7puitsEd46qYIgPLYBoOnPvy7KJO0ZIgwa7gZ0oGLyBxypcpb1lH3ko2GD74G+CjgsnJ7yXENDBZaNwPMsXEFnhw8X5wyy8kme3gvDUVvJ6CxNXyna2nQbS80aQSXapLRvInE0rTvGRCVpzhJizaSuPfOhcwFedH2vrQM4ddrwNvGDPsecZLJRKYCXOMad2OrcVpPUM1n5tBK1Y1Iy8qBVl0jkRNdbJgrnFFzeAbjAOSkob8ufLLrsQm9CJ/EZ9YWYVVL7G+QE6Nnh15hXNrsnSen9Fei23V6KI6MJ6ojP2QdDbHl5RGKLRSUObGhQqkFQZmxqYWyBFevNbZXDlFy/vvj/xyesqxesq6x1HNe0IKSph3gXaaRpntMuv93FLO5U32rGyrbRfc9y2vILyK2q96ZpHufOpk/ozsgOW3cSsKRsNW11BOZNvWClfWuD2YEezQssra+tgTTuamcFp3iV6tEfe0SFipNYVRoxYn0dUlEqo0jLlD02sjaD0EeQs5q/LWBUlaXfB0Iv0MfmtE68GL6lJXuSNnJwZwAahELT1erbRNxOIeb2wrVEkbaQhc2jHvllK+vAkVsVmfYaysAZlF1S7QfkZRoOqqLCSaxeIJwoKb9TLa3JuqUcfZNU+8Q7kZCv6JGlmkXtNNo/5IvrVx394Geon5w05tgtPk5lFxxGRwwbV1QcS4Vl5dJslJi1f4RCSermCRZBl59TMir6DYpvtDygKMIeR6j75XF1ePUVLbKTMlM2HCDPDwImyXJZtmtDN5bMSZnPgIKyR1WjjLDOMs1GTjDNZjbUnicIKj2J7AVwv4iDwYJ9MeGBPCWHzErxK1Osit05pVLFlwc77sgWmqIT3ybFS1WmdqrZLeta+oT20ZTTHy0q7ZC+xBLFFVqIUiSksLqO/Or1fY70bwdLtcl20b+nt0jWv83xwAYdGyZCNT6qw/JOyA2774kxD7mQHqm7Jhol1vtA7RL9tkgYJQfC7Qv+LcpQ+cDxkXXJYxvSw5FCqSw7zYVSCTd8Crgudg4bHSA7NbwuefGXYNxiS7OBBwdJZj8oiQ9gILJ0cLI8h8qN99wcMi8Byo5uEJfEUU7h0e98PGRN2uqgmTJ5Gp2fnb+o9GMCXi9uh/HZ7Py1zXOwP9nHSK0P9wAIeaew7aGxWFRGtOmO6LqByp1eoeaTkmjxyJlCs+J9+m98hApU7BqZtoMq8dpFnoJpcMRaz2GmMnGaj2JXY83qQ2lAukZ9oFKgSu5+4FrKmgDKkvNPXKFD5CAmEaGT3o2zM/LSGA3L/FGl9qFRXAMuEJx8pLeCZ+POzGA4YnibTPz64upihy6evQLMvBt5DZ/ToM71zWoEdq/BXUH5FEleiy/DFSsrwRYmuwy8et7qkLpNzWOHCkXM+xjkCFRMaFXGOQFuFpjjnTR+q96b5/EeZwqcvtzfQ+fZcSufs7rI+/QYx5obxuopUWeTe7yoKUVA5o3eZjW27S2+jwxm5S5TD7zC5O2W2xjdJ7jJvm91dcue2GDRJb2ly8/xv//376+Wg++w8310D9Z8yxnJfqrJUoz3VwqR9Ccu5L7Sn1brS4HZTJu1LhBf7Qnua72vUOfjSR6TqeunPY0pZt8iEpMf/</diagram></mxfile>
