+++ crates/consensus/executor/src/lib.rs
@@ -1,7 +1,13 @@
 // SPDX-License-Identifier: Apache-2.0
 //! Process consensus output and execute every transaction.
 
-#![warn(future_incompatible, nonstandard_style, rust_2018_idioms, rust_2021_compatibility)]
+#![warn(
+    future_incompatible,
+    nonstandard_style,
+    rust_2018_idioms,
+    rust_2021_compatibility,
+    unused_crate_dependencies
+)]
 
 mod errors;
 pub mod subscriber;
