+++ crates/consensus/primary-metrics/src/lib.rs
@@ -1,5 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
 //! Metrics for primary node
+#![warn(
+    future_incompatible,
+    nonstandard_style,
+    rust_2018_idioms,
+    rust_2021_compatibility,
+    unused_crate_dependencies
+)]
 
 pub mod metrics;
 pub use metrics::*;
