--- crates/consensus/primary/README.md
@@ -1,6 +1,0 @@
-# Primary Diagram
-The diagram below illustrates the primary's architecture and could be useful to keep in mind while going through the code.
-
-<p align="center">
-  <img src="../.assets/diagram-primary.svg">
-</p>
