--- crates/consensus/primary/proptest-regressions/tests/certificate_tests.txt
@@ -1,12 +1,0 @@
-# Seeds for failure cases proptest has generated in the past. It is
-# automatically read and these particular cases re-run before any
-# novel cases are generated.
-#
-# It is recommended to check this file in to source control so that
-# everyone who runs the test benefits from these saved cases.
-cc f6ea4ff3eb8f77d909379124536dc2008de434e0696e4f9e93183b57b51651aa # shrinks to num_keys = 6
-cc d66894e197d4e9583b544e78f7b5c32d22410bc5aade4e62767ae7a0404ab42d # shrinks to num_keys = 170
-cc fbd53947e06487d812129cd8431ea890b979ab7b7f5e5920b1b74df01b8421b0 # shrinks to committee_size = 4
-cc 0b8d04af7631183ec6892408476ee07df2ba5e66127f037f0c135bc459c0c101 # shrinks to committee_size = 78
-cc f6e23feae861edc668a8612741f71dc7850d997273f5e31a7d6bc7d96f201b9c # shrinks to committee_size = 4
-cc 3b0f0e02ab5d24061949008979a78c94b0210bb551f8b0eedda25c4cd5e74329 # shrinks to committee_size = 4
