+++ crates/consensus/primary/src/aggregators/mod.rs
@@ -1,6 +1,5 @@
 //! Aggregate messages from peers.
 
 pub(crate) mod certificates;
-pub(crate) mod sync;
 mod votes;
 pub(crate) use votes::VotesAggregator;
