+++ crates/consensus/primary/src/error/cert_manager.rs
@@ -17,13 +17,9 @@ pub(crate) enum CertManagerError {
     /// Error from garbage collection task.
     #[error(transparent)]
     GC(#[from] GarbageCollectorError),
-
     /// The certificate's signature verification state is unverified.
     #[error("Unverified signature verification state {0}")]
     UnverifiedSignature(CertificateDigest),
-    /// The parent was not found.
-    #[error("Error accepting certificate. Parent not found: {0}")]
-    MissingParent(CertificateDigest),
     /// Oneshot channel dropped for certificate manager.
     #[error("Failed to return certificate manager's result.")]
     CertificateManagerOneshot,
