--- crates/consensus/worker/README.md
@@ -1,6 +1,0 @@
-# Worker Diagram
-The diagram below illustrates the worker's architecture and could be useful to keep in mind while going through the code.
-
-<p align="center">
-  <img src="../.assets/diagram-worker.svg">
-</p>
