+++ crates/execution/batch-builder/Cargo.toml
@@ -23,7 +23,6 @@ tracing = { workspace = true }
 # unit tests
 assert_matches = { workspace = true }
 tempfile = { workspace = true }
-eyre = { workspace = true }
 tn-reth = { workspace = true, features = ["test-utils"] }
 
 # integration tests
