+++ crates/execution/batch-validator/src/lib.rs
@@ -1,7 +1,13 @@
 // SPDX-License-Identifier: MIT or Apache-2.0
 //! Batch validation
 
-#![warn(unused_crate_dependencies)]
+#![warn(
+    missing_debug_implementations,
+    missing_docs,
+    unreachable_pub,
+    rustdoc::all,
+    unused_crate_dependencies
+)]
 
 mod validator;
 pub use validator::BatchValidator;
