+++ crates/execution/faucet/Cargo.toml
@@ -41,14 +41,12 @@ secp256k1 = { workspace = true }
 tn-types = { workspace = true }
 
 [dev-dependencies]
-prometheus = { workspace = true }
 consensus-metrics = { workspace = true }
 tn-test-utils = { workspace = true }
 tn-network-types = { workspace = true }
 tn-worker = { workspace = true }
 tn-storage = { workspace = true }
 tempfile = { workspace = true }
-serde_json = { workspace = true }
 tn-config = { workspace = true }
 tn-reth = { workspace = true, features = ["test-utils"] }
 tn-types = { workspace = true, features = ["test-utils"] }
