+++ crates/network-libp2p/src/lib.rs
@@ -6,7 +6,13 @@
     html_favicon_url = "https://www.telco.in/logos/TEL.svg",
     issue_tracker_base_url = "https://github.com/telcoin-association/telcoin-network/issues/"
 )]
-#![warn(missing_debug_implementations, missing_docs, unreachable_pub, rustdoc::all)]
+#![warn(
+    missing_debug_implementations,
+    missing_docs,
+    unreachable_pub,
+    rustdoc::all,
+    unused_crate_dependencies
+)]
 #![deny(unused_must_use, rust_2018_idioms)]
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
