+++ crates/network-types/Cargo.toml
@@ -19,7 +19,6 @@ description = "Commonly used types in telcoin network's internal network communi
 tn-types = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
 tracing = { workspace = true }
-roaring = { workspace = true }
 async-trait = { workspace = true }
 eyre = { workspace = true }
 libp2p = { workspace = true }
