+++ crates/node/Cargo.toml
@@ -17,7 +17,6 @@ thiserror = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 tokio-stream = { workspace = true }
 tracing = { workspace = true }
-url = { version = "2.3.1", optional = true }
 tn-executor = { workspace = true }
 tn-primary = { workspace = true }
 prometheus = { workspace = true }
